Undo refactor out Deck class
